Restructuration de la librairie: class, class utilities, and data loader
Dans l'esprit de séparer les classes d'objets des scripts de chargement de données, déplace "global_fishing_watch.py" dans un dossier "data_loading_scripts" à part.

Suggère également de créer des sous-fichiers type "vessel_utils" pour héberger les méthodes de classes trop grandes afin de bien différencier classes et méthodes.
